Add labels + JMP instruction
This commit adds labels, which are names for code locations, declared
using the following syntax:
```
label:
```

The new JMP instruction jumps to the instruction following the label
(sets ip to be equal to the label's address).

Labels will allow us to do some great stuff in the future, like if/else
and switch statements, as well as calling functions and etc.