Allow variable-parameter instructions
This commit makes us check for a newline inside of the parseInstruction
function. That makes us validate the instruction before returning it from the
function.
For example:
if we had 'push ax 3' we would parse 'push ax' and return it, then
check if there's a newline afterhands. When this didn't work, we would error,
but it could be possible that there's a different rule that makes the syntax
possible.