Add the negative operator
Add the '-' operator so we can create negative numbers or flip the sign.

This also fixes problems with lexing negative numbers, by separating the minus
and the number.
For example: `3-3` didn't work because we lexed "3" and "-3" instead of "3", "-" and "3".
This update allows us to control how we approach such an expression.