Make mov/sub/add/cmp uniform + add more combinations
Finally!!!!

This commit allows us to add binary operations soooooo
much easily.
The commit generalizes binary operations and adds new ones that
work with every operation.
This is a huge step forward towards supporting all of the basic
arithmetic functions (xor, and etc).
Now instead of wasting hours and duplicating tons on tons of code
to create really small instruction combination handlers,
we can now add a new instruction with all of the functionality
we want in just a few minutes.