Fix a problem of not allowing to write to byte-memory
If we wanted to do something like:
mov [byte x] 3
We would actually translate that to:
mov [word x] 3
which has a whole other meaning and is
REALLY bad.
The problem occurred because of us allowing byte memory to be
treated the same as word memory (translating Memory8 to Memory16
when needed). This was a stupid thing to have.

This commit just removes that translation part :).

This also makes dumb stuff like:
mov [byte x] ax
not work.