Fix problem that caused jumps to always go to position 0 in code
An example code for the problem:
```
    mov ax 5
    debug

the_loop:
    mov ax 3
    debug

    jmp the_loop
```

The code above should show that ax equals 5 then loop forever and
show that it's equal to 3, but before the fix we would forever
show that it's 5 then that it's 3.

The problem was caused because we didn't store the right parse rule when we parsed an instruction.
That caused label processing functions (functions that decide what offset
instructions and labels are located in) to not run (we always stored a parse rule that doesn't call those
functions).
That left the offsets uninitialized, and because uninitialized values are commonly equal to 0,
that made the jump addresses 0.